[CSL-1871] Reorganize update system modules in cardano-sl-core
