+++ core/Pos/Core/Block/Main/Lens.hs
@@ -79,8 +79,9 @@ import           Pos.Core.Delegation (DlgPayload)
 import           Pos.Core.Slotting.Types (SlotId)
 import           Pos.Core.Ssc (SscPayload)
 import           Pos.Core.Txp (Tx, TxPayload, TxWitness, txpTxs, txpWitnesses)
-import           Pos.Core.Types (BlockVersion, ChainDifficulty, HeaderHash, SoftwareVersion)
+import           Pos.Core.Types (ChainDifficulty, HeaderHash)
 import           Pos.Core.Update (UpdatePayload)
+import           Pos.Core.Update.Types (BlockVersion, SoftwareVersion)
 import           Pos.Crypto (Hash, PublicKey)
 import           Pos.Merkle (MerkleTree)
