+++ core/Pos/Core/Genesis/Canonical.hs
@@ -33,9 +33,9 @@ import           Pos.Core.Genesis.Types (GenesisAvvmBalances (..), GenesisData (
 import           Pos.Core.Slotting.Types (EpochIndex (..), Timestamp (..))
 import           Pos.Core.Ssc.Types (VssCertificate (..), VssCertificatesMap (..))
 import           Pos.Core.Ssc.Vss (validateVssCertificatesMap)
-import           Pos.Core.Types (Address, BlockVersionData (..), Coin, CoinPortion, SharedSeed (..),
-                                 SoftforkRule (..), StakeholderId, getCoinPortion, mkCoin,
-                                 mkCoinPortion, unsafeGetCoin)
+import           Pos.Core.Types (Address, Coin, CoinPortion, SharedSeed (..), StakeholderId,
+                                 getCoinPortion, mkCoin, mkCoinPortion, unsafeGetCoin)
+import           Pos.Core.Update.Types (BlockVersionData (..), SoftforkRule (..))
 import           Pos.Crypto (ProxyCert, ProxySecretKey (..), PublicKey, RedeemPublicKey, Signature,
                              decodeAbstractHash, fromAvvmPk, fullProxyCertHexF, fullPublicKeyF,
                              fullSignatureHexF, hashHexF, parseFullProxyCert, parseFullPublicKey,
