+++ update/Pos/Update/Download.hs
@@ -28,8 +28,7 @@ import           System.Wlog (WithLogger, logDebug, logInfo, logWarning)
 
 import           Pos.Binary.Class (Raw)
 import           Pos.Binary.Update ()
-import           Pos.Core.Types (SoftwareVersion (..))
-import           Pos.Core.Update (UpdateData (..), UpdateProposal (..))
+import           Pos.Core.Update (SoftwareVersion (..), UpdateData (..), UpdateProposal (..))
 import           Pos.Crypto (Hash, castHash, hash)
 import           Pos.Exception (reportFatalError)
 import           Pos.Reporting (reportOrLogW)
