+++ wallet/src/Pos/Wallet/Aeson/ClientTypes.hs
@@ -7,9 +7,10 @@ import           Universum
 import           Data.Aeson (FromJSON (..), ToJSON (..), object, (.=))
 import           Data.Aeson.TH (defaultOptions, deriveJSON, deriveToJSON)
 import           Data.Version (showVersion)
+import           Servant.API.ContentTypes (NoContent (..))
 
 import           Pos.Client.Txp.Util (InputSelectionPolicy)
-import           Pos.Core.Types (SoftwareVersion (..))
+import           Pos.Core.Update (SoftwareVersion (..))
 import           Pos.Util.BackupPhrase (BackupPhrase)
 import           Pos.Wallet.Aeson.Options (customOptionsWithTag)
 import           Pos.Wallet.Web.ClientTypes (Addr, ApiVersion (..), CAccount, CAccountId,
@@ -22,7 +23,6 @@ import           Pos.Wallet.Web.ClientTypes (Addr, ApiVersion (..), CAccount, CA
                                              SyncProgress, Wal)
 import           Pos.Wallet.Web.Error (WalletError)
 import           Pos.Wallet.Web.Sockets.Types (NotifyEvent)
-import           Servant.API.ContentTypes (NoContent (..))
 
 deriveJSON defaultOptions ''CAccountId
 deriveJSON defaultOptions ''CWAddressMeta
